refactor : goal 타입 분리
