+++ src/types/common.ts
@@ -1,0 +1,3 @@
+export interface regularObject {
+  [key: string]: any;
+}
