+++ src/types/goals.ts
@@ -1,0 +1,16 @@
+import { regularObject } from "./common";
+
+export interface SavedMoney {
+  date: string;
+  money: number;
+}
+
+export interface GoalData extends regularObject {
+  currentMoney?: SavedMoney[];
+  goalDate?: string;
+  goalMoney?: string;
+  goalTitle?: string;
+  userId?: string;
+  userName?: string;
+  id?: string;
+}
